+++ R/normalise.R
@@ -11,6 +11,8 @@
 #' is 400, and it is generally advisable to retain this default setting.
 #' @param offset Integer. Specifies the number of histogram bins to ignore during normalization. This value
 #' will be stored as an attribute in the output object.
+#' @param legacy Logical. Whether to use the legacy C++ code for normalisation (for backward
+#' compatibility) or the modern R version (default). The modern version solves the edge effect issue.
 #' @param filename Char (optional). The output file name for the .text file.
 #'
 #' @seealso \code{\link{histogram}}, and \code{\link{benchmark}}
@@ -30,6 +32,7 @@ normalise <- function(
         bin_width = NULL,
         trim_size = 400,
         offset = 0,
+        legacy = FALSE,
         filename = "") {
 
     # check bin_width and get it from histo object
@@ -43,12 +46,24 @@ normalise <- function(
         }
     }
 
+    # remove the 0,0 point
+    nr <- nrow(x)
+    x[nr, 1] <- 0
+
     # normalise the histo C++
-    out <- norm_cpp(
-        x = x,
-        trim_size = trim_size,
-        offset = offset
-    )
+    if (legacy) {
+        out <- norm_cpp(
+            x = x,
+            trim_size = trim_size,
+            offset = offset
+        )
+    } else {
+        out <- norm_r(
+            x = x,
+            trim_size = trim_size,
+            offset = offset
+        )
+    }
 
     # write down the histogram
     if (nchar(filename) > 0) {
@@ -79,6 +94,33 @@ normalise <- function(
 
     return(
         out
+    )
+}
+
+# the new normalisation function that doesn't remove the edges
+norm_r <- function(x, trim_size = 400, offset = 0) {
+    nr <- nrow(x)
+    x[nr, 1] <- 0
+    r <- terra::rast(x)
+    # create a Gaussian filter and normalise it to add up to 1
+    d <- stats::dnorm(-2:2, 0, 1)
+    w <- outer(d, d)
+    w <- w / sum(w)
+    rr <- terra::focal(r, w = w, fun = sum, na.rm = TRUE)
+    mat <- as.matrix(rr, wide = TRUE)
+
+    # trim the histogram and apply offset
+    lower <- nr - trim_size + 1 - offset
+    upper <- nr - offset
+    left <- 1 + offset
+    right <- trim_size + offset
+    mat <- mat[lower:upper, left:right]
+
+    # normalise the columns and then reverse them
+    mat <- apply(mat, 2, FUN = function(x) rev(x / sum(x)))
+
+    return(
+        t(mat)
     )
 }