+++ NEWS.md
@@ -1,6 +1,8 @@
 # Version 0.2.0
 * Added `drop_features` parameter to fully exclude specific variables from both the `histogram` and `benchmark` functions.
 * Adding a condition to ensure that `k_obs`  is less than or equal to `k_pred`.
+* A new histogram normalisation is added using R code (the `legacy = FALSE` argument) to remove edge effect.
+* The zero-zero point in the raw histogram (self-assessment) is removed before normalisation.
 
 # Version 0.1.6
 * Added a `scale_factor` parameter to allow user-defined correction of geographic CRS distance calculations, enhancing flexibility in handling distance conversions (previously set to a default value only).